chore(dogstatsd): extract well-known DSD tags to dedicated struct + handle host tag
