+++ lib/saluki-components/src/sources/dogstatsd/origin.rs
@@ -193,7 +193,7 @@ impl OriginTagsResolver for DogStatsDOriginTagResolver {
 /// Builds an `RawOrigin` object from the given metric packet.
 pub fn origin_from_metric_packet<'packet>(packet: &MetricPacket<'packet>) -> RawOrigin<'packet> {
     let mut origin = RawOrigin::default();
-    origin.set_pod_uid(packet.pod_uid);
+    origin.set_pod_uid(packet.well_known_tags.pod_uid);
     origin.set_container_id(packet.container_id);
     origin.set_external_data(packet.external_data);
     origin.set_cardinality(packet.cardinality);
@@ -203,7 +203,7 @@ pub fn origin_from_metric_packet<'packet>(packet: &MetricPacket<'packet>) -> Raw
 /// Builds an `RawOrigin` object from the given event packet.
 pub fn origin_from_event_packet<'packet>(packet: &EventPacket<'packet>) -> RawOrigin<'packet> {
     let mut origin = RawOrigin::default();
-    origin.set_pod_uid(packet.pod_uid);
+    origin.set_pod_uid(packet.well_known_tags.pod_uid);
     origin.set_container_id(packet.container_id);
     origin.set_external_data(packet.external_data);
     origin.set_cardinality(packet.cardinality);
@@ -213,7 +213,7 @@ pub fn origin_from_event_packet<'packet>(packet: &EventPacket<'packet>) -> RawOr
 /// Builds an `RawOrigin` object from the given service check packet.
 pub fn origin_from_service_check_packet<'packet>(packet: &ServiceCheckPacket<'packet>) -> RawOrigin<'packet> {
     let mut origin = RawOrigin::default();
-    origin.set_pod_uid(packet.pod_uid);
+    origin.set_pod_uid(packet.well_known_tags.pod_uid);
     origin.set_container_id(packet.container_id);
     origin.set_external_data(packet.external_data);
     origin.set_cardinality(packet.cardinality);
